Change location of header files.
They where moved from manifold_learning to
manifold_learning/src/. Also changed the build scripts to build in the
same place all the extensions that depend on these headers.

It migh not be a good practice to build extension modules from
submodules, but in this case we have 3 submodules that share the same
dependencies and the alternative is to repeat the same code in 3
places. Anyway, in the future a better structure for this directory
should be envisaged. For this release, let's leave it like this.

As a side-effect, the $PATH issue in OsX with boost installed
from MacPorts dissapeared, probably because passing
numpy.get_include() to include_dirs adds the needed dirs (if numpy is
installed through MacPorts).

Also corrected some setup.py scripts that where throwing
error/warnings.

git-svn-id: https://scikit-learn.svn.sourceforge.net/svnroot/scikit-learn/trunk@362 22fbfee3-77ab-4535-9bad-27d1bd3bc7d8