Hello list,
I'd like to commit a new feature to the scikits-learn project.
I implemented a kriging model class which is able to perform both regression
and probabilistic classification.
I will send an e-mail with a script that performs a demo.
Since I have no clue where to put my contribution I simply created a single
file named kriging.py with the class and other functions such as correlation
and regression models, and I added an import instruction to the main
__init__.py module of scikit.learn.

I hope you'll enjoy this contribution.

Vincent