Separate correction and reweighting steps from raw MCD computation.
Thus, it is possible to run only once the FastMCD algorithm and to
correct/reweight from the raw estimates, resulting in a computation
time gain. Moreover, its possible to compare different correction or
reweighting methods on the basis of the same raw estimates.
The correction/reweigthing can be made permanent and since the raw
estimates are stored in the MinCovDet object anyway, one can come back
to them if needed.