Integrate Gael's minor comments + Magnify examples + 1D data case.
Review of the variables names suggested by Gael.
Better illustation on examples (boxplot, optimized ranges).
One-dimensional data are now handled in a way that is consistent with
the rest of sklearn: arrays of shape (x, ) are considered as a sample
with only one observation described by x features and a warning suggests
to the user that he should reshape his array. This fix an issue opened
in the issues list.