optimisations to Ridge Regression GCV
Optimisations follow from noting that we do
not need to explicitly form the matrix for
the regularised inverse, we just need to
compute its action upon the response vector
y. Similarly, we don't need to compute the
whole explicit matrix in order to evaluate
its diagonal, instead just evaluate the
diagonal entries.

This leads to a large performance increase
when GCV is applied to training datasets
with larger numbers of points. E.g. I observe
a ~ 11x speedup on the UCI Communities and
Crime dataset even for modest training set
sizes of 398 examples only (20% of the data).