MAINT refactor LSH forest code
Removed _do_hash function.
The functionality of _do_hash is done within the _create_tree function as
a hashing function is generated for each tree and it doesn't need a separate
function.

Added _convert_to_hash function.
This will be used in the queries and insertions to convert the data point
into the integer represented by the binary hash.

Updated _generate_masks to work with cached hashes.

Changed the insert function to accpet a batch of data points(a 2D array).
Now trees and original indices are stored in lists. Tests for insert and fit
functions have been updated.

Changed safe_asarray to check_arrary.