MAINT make it possible to use wheelhouse-uploader
wheelhouse-uploader makes it possible to leverage our CI infrastructure
(travis and appveyor) to automate the generation of binary packages
for OSX and windows (both .exe and .whl packages).

The cloud containers that are used to store the build artifacts of the
CI workers are configured in the setup.cfg file.

More details at: https://github.com/ogrisel/wheelhouse-uploader